Crate crates_index

Library for retrieving and interacting with the crates.io index.

Examples

let index = crates_index::Index::new_cargo_default().unwrap();
if !index.exists() {
   index.retrieve().expect("Could not fetch crates.io index");
}
for crate_ in index.crates() {
   let latest_version = crate_.latest_version();
   println!("crate name: {}", latest_version.name());
   println!("crate version: {}", latest_version.version());
}

Structs

Crate

A single crate that contains many published versions

Crates

Iterator over all crates in the index. Skips crates that failed to parse.

Dependency

A single dependency of a specific crate version

Index

Wrapper around managing the crates.io-index git repository

IndexConfig

Global configuration of an index, reflecting the contents of config.json as specified at https://doc.rust-lang.org/cargo/reference/registries.html#index-format

Version

A single version of a crate published to the index

Enums

DependencyKind

Error