Struct crates_index::Index

pub struct Index {
    // some fields omitted
}

Wrapper around managing the crates.io-index git repository

Methods

impl Index

fn new(path: PathBuf) -> Index

Construct a new Index supplying a path where the index lives or should live

fn exists(&self) -> bool

Determines if anything exists at the path specified from the constructor

fn clone(&self) -> Result<(), Error>

Clones the index to the path specified from the constructor

fn crate_(&self, crate_name: &str) -> Option<Crate>

Retrieve a single crate by name (case insensitive) from the index

fn crates(&self) -> Crates

Retrieve an iterator over all the crates in the index

fn crate_index_paths(&self) -> CrateIndexPaths

Returns all the crate index file paths in the index