Crate crates_index

Library for retrieving and interacting with the crates.io index.

Examples

extern crate crates_index;

let index = crates_index::Index::new("_index");
if !index.exists() {
   index.retrieve().expect("Could not fetch crates.io index");
}
for crate_ in index.crates() {
   let latest_version = crate_.latest_version();
   println!("crate name: {}", latest_version.name());
   println!("crate version: {}", latest_version.version());
}

Structs

Crate
A single crate that contains many published versions

CrateIndexPaths
Constructed from Index::crate_index_paths

Crates
Constructed from Index::crates

Dependency
A single dependency of a specific crate version

Error
The Error type.

Index
Wrapper around managing the crates.io-index git repository

Version
A single version of a crate published to the index

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.