Struct crates_index::Index

pub struct Index { /* fields omitted */ }

Wrapper around managing the crates.io-index git repository

Methods

impl Index

pub fn new<P: Into<PathBuf>>(path: P) -> Index

Construct a new Index supplying a path where the index lives or should live

pub fn exists(&self) -> bool

Determines if a crates.io repository exists at self.path

pub fn retrieve(&self) -> Result<()>

Downloads the index to the path specified from the constructor

pub fn update(&self) -> Result<()>

Assumes the index already exists at self.path, and updates it

pub fn retrieve_or_update(&self) -> Result<()>

Downloads the index to the path specified from the constructor

pub fn crate_(&self, crate_name: &str) -> Option<Crate>

Retrieve a single crate by name (case insensitive) from the index

pub fn crates(&self) -> Crates

Retrieve an iterator over all the crates in the index

pub fn crate_index_paths(&self) -> CrateIndexPaths

Returns all the crate index file paths in the index

pub fn path(&self) -> &Path

Get the index directory.